requirements.txt
# --- Flask Core ---
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
click==8.1.7

# --- PostgreSQL ---
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32

# --- Configuración ---
python-dotenv==1.0.1

# --- Reconocimiento Facial (solo PyTorch, sin TensorFlow) ---
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.10.0.84
deepface==0.0.93
retina-face==0.0.14
torch==2.3.1
torchvision==0.18.1
# tf-keras se mantiene liviano y evita conflictos internos
tf-keras==2.18.0

# --- Render Optimization ---
# No instalamos mtcnn ni tensorflow pesado
